Move cfgstruct/utils to cmd/internal/wizard
